API/central_agent.py
#### RUN THIS TO INSTALL PKGS ########
# This comment indicates that the following line is a shell command to install necessary packages
# python -m pip install semantic-kernel
######################################

# Import necessary packages
import semantic_kernel as sk  # Import the semantic_kernel library
from semantic_kernel import PromptTemplateConfig, PromptTemplate, SemanticFunctionConfig  # Import additional classes from semantic_kernel
from sklearn.feature_extraction.text import TfidfVectorizer  # Import TfidfVectorizer for text vectorization
from sklearn.metrics.pairwise import cosine_similarity  # Import cosine_similarity for computing similarity between texts
import random  # Import random module for generating random numbers
import time  # Import time module for time-related tasks
import asyncio

# Define a function to calculate cosine similarity between two texts
def calculate_cosine_similarity(text1, text2):
    vectorizer = TfidfVectorizer()  # Initialize a TfidfVectorizer
    tfidf_matrix = vectorizer.fit_transform([text1, text2])  # Transform the texts into TF-IDF matrix
    return cosine_similarity(tfidf_matrix[0:1], tfidf_matrix[1:2])[0][0]  # Return the cosine similarity score

# Define a function to load text lines from a file
def load(filename):
    lines = []  # Initialize an empty list to store lines
    with open(filename, 'r') as file:  # Open the file in read mode
        while True:
            line = file.readline()  # Read a line from the file
            if not line:  # Break the loop if no more lines
                break
            lines.append(line.strip())  # Add the line to the list after stripping whitespace
    return '\n'.join(lines)  # Return the lines joined by newlines

# Define a class to represent a Central Agent
class CentralAgent:
    def __init__(self):
        self.input_text = None  # Initialize input_text attribute to None
        self.kernel = sk.Kernel()  # Initialize a semantic kernel

    # Method to upload input text
    def upload_input_text(self, input_text):
        self.input_text = input_text  # Set the input_text attribute

    # Async method to assign tasks to different Nodes
    async def assign_tasks(self):
        if self.input_text:
            # Create and execute a semantic function to give a response 
            return self.kernel.create_semantic_function(f"""Break down the tasks that are given in the {self.input_text} and assign them to specific agents based on the functionality and what is supposed to be done with them. Make sure to provide clear instructions to the agents for the specified task""", max_tokens=1024)()

# Define a class to represent a Node Agent


# Define a class to represent a General Agent
class GeneralAgent:
    def __init__(self):
        self.kernel = sk.Kernel()  # Initialize a semantic kernel
    
# Define a coroutine to assign tasks to different agents
async def assign_tasks(central_agent, node_agent):
    central_agent.upload_input_text(f"""""")  # Upload input text

    action = await central_agent.assign_tasks()  # Assign tasks to each agent





# Execute the main function if this script is run as the main module
if __name__ == "__main__":
    import asyncio  # Import asyncio for asynchronous execution
    asyncio.run()  
