requirements.txt
autogen==0.3.1
fastapi==0.115.2
google_api_python_client==2.147.0
google_auth_oauthlib==1.2.1
huggingface_hub==0.23.4
langchain_huggingface==0.1.0
protobuf==3.20.3
python-dotenv==1.0.1
scikit_learn==1.5.2
semantic_kernel==1.11.0
transformers==4.41.2
